+++ bluetooth_low_energy_platform_interface/lib/src/my_central.dart
@@ -1,13 +1,11 @@
 import 'central.dart';
-import 'my_object.dart';
 import 'uuid.dart';
 
-class MyCentral extends MyObject implements Central {
+class MyCentral implements Central {
   @override
   final UUID uuid;
 
   MyCentral({
-    required super.hashCode,
     required this.uuid,
   });
 }