+++ bluetooth_low_energy_platform_interface/lib/src/my_gatt_characteristic.dart
@@ -1,10 +1,9 @@
 import 'gatt_characteristic.dart';
 import 'gatt_characteristic_property.dart';
 import 'my_gatt_descriptor.dart';
-import 'my_object.dart';
 import 'uuid.dart';
 
-class MyGattCharacteristic extends MyObject implements GattCharacteristic {
+class MyGattCharacteristic implements GattCharacteristic {
   @override
   final UUID uuid;
   @override
@@ -13,7 +12,6 @@ class MyGattCharacteristic extends MyObject implements GattCharacteristic {
   final List<MyGattDescriptor> descriptors;
 
   MyGattCharacteristic({
-    super.hashCode,
     required this.uuid,
     required this.properties,
     required this.descriptors,