+++ bluetooth_low_energy_platform_interface/lib/src/my_peripheral_manager.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/cupertino.dart';
 import 'package:log_service/log_service.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
@@ -7,7 +6,7 @@ import 'peripheral_manager.dart';
 /// Platform-specific implementations should implement this class to support
 /// [PeripheralManager].
 abstract class MyPeripheralManager extends PlatformInterface
-    with LoggerController
+    with LoggerProvider, LoggerController
     implements PeripheralManager {
   static final Object _token = Object();
 
@@ -34,8 +33,4 @@ abstract class MyPeripheralManager extends PlatformInterface
 
   /// Constructs a [MyPeripheralManager].
   MyPeripheralManager() : super(token: _token);
-
-  @protected
-  @override
-  Logger get logger => Logger('$PeripheralManager');
 }